feat: api 통신을 위한 상태 및 함수 선언 위치 변경
